Fixes for better compilation on Mac
